+++ app/core/utils.coffee
@@ -119,8 +119,8 @@ courseIDs =
   COMPUTER_SCIENCE_5: '569ed916efa72b0ced971447'
   COMPUTER_SCIENCE_6: '5817d673e85d1220db624ca4'
 
-# TODO add when final courses content created for voyager 
-voyagerCourseIDs = []
+# TODO add when final courses content created for ozaria 
+ozariaCourseIDs = []
 
 orderedCourseIDs = [
   courseIDs.INTRODUCTION_TO_COMPUTER_SCIENCE
@@ -782,5 +782,5 @@ module.exports = {
   premiumContent
   isValidEmail
   videoLevels
-  voyagerCourseIDs
+  ozariaCourseIDs
 }
