+++ app/schemas/models/campaign.schema.coffee
@@ -78,7 +78,7 @@ _.extend CampaignSchema.properties, {
       #- normal properties
       position: c.point2d()
 
-      # properties relevant for voyager campaigns
+      # properties relevant for ozaria campaigns
       nextLevels: {
         type: 'array'
         description: 'array of next levels original id'
