+++ app/views/play/CampaignView.coffee
@@ -110,12 +110,12 @@ module.exports = class CampaignView extends RootView
     @terrain = 'picoctf' if window.serverConfig.picoCTF
     @editorMode = options?.editorMode
     @requiresSubscription = not me.isPremium()
-    # Allow only admins to view the voyager campaign and only in editor mode
-    # New page for non-editor mode `/play-voyager`
-    # Assuming, the voyager placeholder campaigns will start with 'voyager'
-    # TODO: Remove/update this check before final voyager launch
-    if _.string.startsWith(@terrain, "voyager") and (not me.showVoyagerCampaign() or not @editorMode)
-      console.error("voyager dummy campaign, only editor mode is available for admins!")
+    # Allow only admins to view the ozaria campaign and only in editor mode
+    # New page for non-editor mode `/play-ozaria`
+    # Assuming, the ozaria placeholder campaigns will start with 'ozaria'
+    # TODO: Remove/update this check before final ozaria launch
+    if _.string.startsWith(@terrain, "ozaria") and (not me.showOzariaCampaign() or not @editorMode)
+      console.error("ozaria dummy campaign, only editor mode is available for admins!")
       return
     if @editorMode
       @terrain ?= 'dungeon'
